+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Fix error parsing Helm version (https://github.com/pulumi/pulumi-kubernetes/pull/1170)
 -   Fix prometheus-operator test to wait for the CRD to be ready before use (https://github.com/pulumi/pulumi-kubernetes/pull/1172)
 -   Fix suppress deprecation warnings flag (https://github.com/pulumi/pulumi-kubernetes/pull/1189)
+-   Set additionalSecretOutputs on Secret data fields (https://github.com/pulumi/pulumi-kubernetes/pull/1194)
 
 ### Improvements